+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- Added `auto_correct` option.
+
 ## 0.1.0 - 2022-08-31
 
 ### Added